requirements.txt
requests==2.31.0
sqlite3==3.0.0
python-telegram-bot==21.1.1